An audio and video viewing and sharing website
The main source of video sharing in the internet is the YouTube system.
Even though there are many video sharing websites YouTube is considered
to be the best of all. But the only limitation we can in YouTube is that
we can upload only videos files. Even though it’s not a drawback, we
thought why not build a system that can support audio, video and images
as well. So we propose our system MeTube, which can support all the
above three media.

This system follows most of the functionalities of YouTube and in
addition it has features to support all media and a place where users
can share their thoughts and discuss about their favorite media i.e. a
forum for them to discuss all about them. Unlike YouTube users in MeTube
can add/remove friends and block people from viewing their media.

The main intention was to build a system which is user-friendly such
that people with less knowledge about the internet world can access and
enjoy our system.